pyproject.toml
[project]
name = "fontconvert"
version = "0.1.0"
description = "UFO-based font build pipeline from SVGs (ASCII default, monospace-first)"
requires-python = ">=3.10"
dependencies = [
  "fontmake>=3.7.0",
  "ufoLib2>=0.16.0",
  "pyyaml>=6.0.0",
  "Pillow>=9.0.0",
]

[project.scripts]
fontconvert = "fontconvert.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fontconvert = ["*.tsv"]
